Added an instructions page.
Implemented NavigationService to handle the settings and instructions views, and reworked all the existing Settings popup stuff to use it.

Fixed a but in the Humanizers where they didn't account for WPF passing in empty strings (I suppose, cause the values are bound to combo boxes, it makes sense that the default value would be an empty string, even if the textbox portion of the control is disabled).